+++ LucyEngine/Source/Renderer/Memory/Buffer/IndexBuffer.h
@@ -9,14 +9,14 @@ namespace Lucy {
 
 	class IndexBuffer : public Buffer<uint32_t> {
 	public:
-		static Ref<IndexBuffer> Create(uint32_t size);
+		static Ref<IndexBuffer> Create(size_t size);
 
 		virtual ~IndexBuffer() = default;
 
 		virtual void LoadToGPU() = 0;
 		virtual void DestroyHandle() = 0;
 	protected:
-		IndexBuffer(uint32_t size);
+		IndexBuffer(size_t size);
 	};
 }
