+++ LucyEngine/Source/Renderer/Memory/Buffer/VertexBuffer.h
@@ -9,14 +9,14 @@ namespace Lucy {
 
 	class VertexBuffer : public Buffer<float> {
 	public:
-		static Ref<VertexBuffer> Create(uint32_t size);
+		static Ref<VertexBuffer> Create(size_t size);
 		
 		virtual ~VertexBuffer() = default;
 
 		virtual void LoadToGPU() = 0;
 		virtual void DestroyHandle() = 0;
 	protected:
-		VertexBuffer(uint32_t size);
+		VertexBuffer(size_t size);
 	};
 }
