+++ LucyEngine/Source/Renderer/Memory/Buffer/Vulkan/VulkanIndexBuffer.h
@@ -11,14 +11,14 @@ namespace Lucy {
 
 	class VulkanIndexBuffer : public IndexBuffer {
 	public:
-		VulkanIndexBuffer(uint32_t size);
+		VulkanIndexBuffer(size_t size);
 		virtual ~VulkanIndexBuffer() = default;
 		
 		void Bind(const VulkanIndexBindInfo& info);
 		void LoadToGPU() final override;
 		void DestroyHandle() final override;
 	private:
-		void Create(uint32_t size = 0);
+		void Create(size_t size = 0);
 
 		VkBuffer m_BufferHandle = VK_NULL_HANDLE;
 		VmaAllocation m_BufferVma = VK_NULL_HANDLE;
