+++ LucyEngine/Source/Renderer/Memory/Buffer/Vulkan/VulkanUniformBuffer.cpp
@@ -56,7 +56,7 @@ namespace Lucy {
 		imageInfo.sampler = image->GetImageView().GetSampler();
 
 		m_ImageInfos.push_back(imageInfo);
-		return m_ImageInfos.size() - 1;
+		return (uint32_t)m_ImageInfos.size() - 1;
 	}
 
 	uint32_t VulkanUniformImageBuffer::BindImage(VkImageView imageView, VkImageLayout layout, VkSampler sampler) {
@@ -66,7 +66,7 @@ namespace Lucy {
 		imageInfo.sampler = sampler;
 
 		m_ImageInfos.push_back(imageInfo);
-		return m_ImageInfos.size() - 1;
+		return (uint32_t)m_ImageInfos.size() - 1;
 	}
 
 	void VulkanUniformImageBuffer::Clear() {
